refactor: CMDP-Rotations-Timeout ohne Float berechnen

+++ 5_System/src/CommandRunner/CommandRunner.h
@@ -63,7 +63,7 @@ private:
     void updatePathProgress();
 
     uint32_t calcPathTimeoutMs(uint16_t targetCm, float speedCms) const;
-    uint32_t calcAngleTimeoutMs(uint16_t targetDeg, float speedDegS) const;
+    uint32_t calcAngleTimeoutMs(uint16_t targetDeg, uint16_t speedDegS) const;
 
     void stopAll();
